Don't relaunch a program if user chooses "not now"
Current behavior will "launch" a program when the user selects "not now", even though the program was never closed, which brings a program the user may not be working in to the front. This change checks for that error "10", and if so, it will not run the reopenCLosedProcess function. The end result is if a user selects not now, that program is not brought to the front.