Update functions.sh
This is cleaner than my earlier commit. appClosed is set to 1 before the app is actually closed. If we defer, set appClosed to something other than 1. The function that calls to reopen the app checks if it is set to 1, so setting it to 0 prevents the app from opening, and will appropriately log " App not closed, so no reopen."