continue running the script even if management action hangs
I've seen this happen in several logs
```
2022-12-03 09:13:11 : INFO  : firefoxpkg : notifying
2022-12-13 08:18:36 : INFO  : firefoxpkg : App not closed, so no reopen.
2022-12-13 08:18:36 : REQ   : firefoxpkg : All done!
2022-12-13 08:18:36 : REQ   : firefoxpkg : ################## End Installomator, exit code 0
```
Installomator notifies that the app is installed, and just hangs there, preventing any further jamf policies from running. In this example, 10 days later I killed "Management Action" and it proceeded to finish the script. Adding the `&` when Management Action is called should prevent this possible hang.