Add "valueFromJSON" to eventually replace "getJSONValue"
This "valueFromJSON" function is my latest "json_value" (Version 2023.2.26-1) function (https://github.com/RandomApplications/JSON-Shell-Tools-for-macOS/blob/main/json_value.sh) but renamed to match the Installomator function naming style.

It is intentionally NOT yet replacing the existing "getJSONValue" function because both need to exist simultaneously until all labels have been updated to use the new "valueFromJSON" function. I plan to work through those and submit PRs for each label that is currently using "getJSONValue", and then will submit a PR to remove "getJSONValue" when Installomator and all labels have been updated and merged. I think giving this function a new name is graceful way to handle this transition while also still having a function name that is intuitive and fits in with other functions in Installomator.

This new "valueFromJSON" function has a handful of benefits over the existing "getJSONValue" function and the only drawback is it's a bit more code, and retrieving nested values is not backwards compatible with the current "getJSONValue" function.

Some of the benefits are of "valueFromJSON" are:

- Can accept stdin so that is can be used like "curl ... | valueFromJSON ..." which can make the code in some labels cleaner and simpler.

- Can accept JSON as the first or last argument so it's flexible for different users styles in labels.

- Does not parse any key or index arguments passed to it directly has JXA code like the current "getJSONValue" function does. This means advanced JavaScript methods cannot be used in "valueFromJSON", but that is not currently being taken advantage of in Installomator and should not be needed for basic version or URL retrievals from JSON, and the way "valueFromJSON" handles arguments is safer overall.

- No knowledge of JavaScript dot or bracket notation is required to use "valueFromJSON" like it is for "getJSONValue" to retrieve nested values or complex key name. To retrieve a value value from a dictionary that is nested in an array with "getJSONValue" the expression "getJSONValue '[0].keyName'" needs to be passed as a single argument. But, with "valueFromJSON" each index or key is simply passed as its own argument like "valueFromJSON '0' 'keyName'". This also means there are no possible characters that could exist in valid key names that would need to be escaped or placed in bracket notation to properly retrieve a key. One example would be a period in a key would need to be retrieved with bracket notation like "getJSONValue '["key.name"]'" which could be confusing for some users when "getJSONValue 'key.name'" does not work as expected. With "valueFromJSON", using "valueFromJSON 'key.name'" would properly retrieve the key with the literal name "key.name".

- Negative array indexes can specified be used when someone needs to retrieve the last value of an array.

- "valueFromJSON" has much better error handling and would output an error to stderr with the name of the key or index that was not retrieved rather than outputting nothing like "getJSONValue" would in some cases.

- Detailed help information is documented at https://github.com/RandomApplications/JSON-Shell-Tools-for-macOS/blob/main/json_value.sh and this URL is commented in the function which may be helpful for users who haven't worked with JSON parsing tools before.