Pass variables as args to all "osascript" commands
Instead of placing variables directly in AppleScript code passed to  "osascript", pass the variables as arguments to be retrieved from within the AppleScript code instead.

To be able to retrieve arguments passed to "osascript" within the AppleScript code, the AppleScript code must be within an "run" function/handler, which is shown and described near the top of the "DESCRIPTION" section in "man osascript".

When a variable is placed directly in AppleScript code, its value is expanded by the shell before being passed to "osascript" and is therefore being interpreted as raw AppleScript code. This means that any special characters such as double quotes (") would need to be manually escaped within the variables contents to not break the quoted AppleScript string that the variable is being placed within.

But, when the values are passed to "osascript" as arguments and then retrieved within the AppleScript code from the arguments array, they are always static string values that are never interpreted as AppleScript code. This means that any literal quotes within a variable are properly displayed without needing to be escaped and without ever breaking the AppleScript code.

Double quotes are just one example, but in general passing variables as arguments prevents any possible arbitrary AppleScript code execution. While that is not really likely with how these function are used in Installomator, passing arguments to other environments that interpret code instead of placing variables directly in that code is a safe practice in general and helps avoids any possible mistakes or bugs in the future.

Also, I have slightly adjusted the behavior of "displaydialog" function to set the "Not Now" button as the dialogs "cancel button" which means the Escape key can now be used to dismiss the alert as is normal and expected in standard macOS dialogs. The adjustments needed for this behavior to not require other changes throughout the code is documented in the "NOTE" comment at the end of the "displaydialog" function.

Finally, I added a new "quitApp" function instead of repeating the same "osascript" command 4 times to quit apps in different case statements within the "checkRunningProcesses" function.

These may be a lot of changes in a single PR, but they were all pretty interconnected. If any part of all these changes is not desired, I can make changes as needed.