Updated Label: GraphPad Prism 9
GraphPad Prism 10 has been released (https://www.graphpad.com/updates/prism-1000-release-notes) and this has caused the current appNewVersion variable for the GraphPad Prism 9 label to show version 10.0.0 as the latest version, which is a new major version. It's unlikely that Prism 9 will receive any more updates so I've changed the appNewVersion label to statically be the last released version of Prism 9 (9.5.1) and also created a PR for a Prism 10 label.