Update nextcloud Label
It looks like starting version 3.9 nextcloud is not publishing the releases on their main nextcloud desktop repo anymore but have a separate one for releases only, but this one is not in the nextcloud organisation on github. but i have checked the packages, signatures and developer ID / notorization and the github commits are signed by a valid and nextcloud org person.