Improved handling of msupdate errors and no updates available
If `msupdate` cannot talk to MAU, `updateTool` will not be used.

`msupdate` will give an “XPC”-error when a MDM solution have not been allowed to initiate communication between `msupdate` and MAU.

Here is the solution for Jamf Pro: https://github.com/pbowden-msft/MobileConfigs/tree/master/Jamf-MSUpdate

Curretnly we have me and a collegue have tested this in Mosyle and Addigy, and even if we deploy similar profiles for them, `msupdate` still gives an error. A workaround might be to stopp the LaunchAgent and start it again, and then it should run, but that is also not pretty, so then Installomator can handle the update if any are available.