Update nordvpn.sh
Removed appName and blockingProcesses because they were set to the default values.

Also, xpath is aliased within Installomator.sh to include the `-e` option when needed. This is for backwards compatibility with older operating systems. Your `appNewVersion` code works when run in terminal, but is coming back empty when run within the context of Installomator.sh