Update Microsoft Endpoint Manager (Intune) icon
The Intune Agent is only available on devices if there are scripts or pkg installs assigned to run. Likely the agent is there most of the time, but there could be cases where it is not present. Especially if some other process other than the Intune agent is running Installomator and wanting to use the icon.