Add functionality to load external label files
Added code to look for and load external label files. I.E. labels that are not embedded into the script. These can be labels that have PRs submitted, but not yet approved into a build release, or private labels that people have created for their own use.

These labels should be saved to this directory:
"/usr/local/Installomator/managed-labels"

Ownership of the directory and the label file are both checked to insure that they are root:wheel owned to prevent potential tempering with the code contained in the labels.

There is also code (that is currently commented out), that would be usable for implementing a separate Installomator signed daily/weekly/monthly (some interval) label only pkg update that would separate the imbedded script labels and potentially place labels into this path:
"/usr/local/Installomator/labels"

Code to download, install and verify the signed pkg with the label updates would need to be added if this were to be implemented.

This is all based on the comments in this issue request:
https://github.com/Installomator/Installomator/issues/1999