Update main.sh
Fixed the comment filters. Thankfully all labels with a comment on the end of the line match the `awk` filter of ' # ', and everywhere else a '#' is used inside a label, it does not have a space on both sides of the '#'. And using grep to better identify comment lines.

From a coding perspective, still not the best for line ending comments, but searching online for ideas only proved how hard it is to do.