Update cyberduck.sh
Horribly bad form to download the same file twice, because 1. it's very unfriendly to NAT'd environments as the server might see the extra hit as cause for alarm, and 2. there is the tenet if chances the information could change between the two downloads, resulting in the downloadURL containing a link to an older version of the app than the version that then gets identified in appNewVersion.