New Release Notes URL
Fixed the releaseNotesURL to the new URL, and changed the appNewVersion to properly check the new release Notes since it is now a list and not a table. It will filter out the Windows Only and Optional updates, and grab the (Mac) version if there is a difference between (Mac) and (Win).